+++ lib/storex/store.ex
@@ -2,7 +2,11 @@ defmodule Storex.Store do
   @doc """
   Called when store session starts.
   """
-  @callback init(session_id :: binary(), params :: any()) :: any()
+  @callback init(session_id :: binary(), params :: any()) ::
+              {:ok, state :: any()}
+              | {:ok, state :: any(), key :: binary()}
+              | {:error, reason :: binary()}
+
   @callback mutation(
               name :: binary(),
               data :: any(),
@@ -34,9 +38,7 @@ defmodule Storex.Store do
 
         @store unquote(env.module)
 
-        def init({params, session}) do
-          init_state = @store.init(session, params)
-
+        def init({session, init_state, params}) do
           {:ok,
            %{
              state: init_state,
@@ -46,9 +48,14 @@ defmodule Storex.Store do
         end
 
         def start_link([], session: session, store: store, params: params) do
-          GenServer.start_link(Server, {params, session},
-            name: Storex.Supervisor.name(session, store)
-          )
+          opts = [name: Storex.Supervisor.name(session, store)]
+
+          with {:ok, state, key} <- init_store(session, params),
+               {:ok, pid} <- GenServer.start_link(Server, {session, state, params}, opts) do
+            {:ok, pid, %{session: session, key: key}}
+          else
+            {:error, reason} -> {:error, reason}
+          end
         end
 
         def handle_cast(:session_ended, state) do
@@ -93,6 +100,23 @@ defmodule Storex.Store do
 
         def handle_call(call, _, _state) do
           raise "Not handled call: #{inspect(call)}"
+        end
+
+        defp init_store(session, params) do
+          @store.init(session, params)
+          |> case do
+            {:ok, state} ->
+              {:ok, state, nil}
+
+            {:ok, state, key} ->
+              {:ok, state, key}
+
+            {:error, reason} ->
+              {:error, reason}
+
+            _ ->
+              raise "Return value of store init should be {:ok, state}, {:ok, state, key} or {:error, reason}"
+          end
         end
       end
     end