+++ priv/static/storex.d.ts
@@ -1,12 +1,14 @@
 declare module "storex" {
-    interface StoreConfig {
+    interface StoreConfig<T> {
         session?: string;
         store: string;
         params: {
             [key: string]: any;
         };
-        subscribe?: () => void;
-        connection?: () => void;
+        subscribe?: (state: T) => void;
+        onConnected?: () => void;
+        onError?: (error: unknown) => void;
+        onDisconnected?: (event: CloseEvent) => void;
     }
     class Storex<T> {
         private session;
@@ -20,13 +22,12 @@ declare module "storex" {
             };
             address?: string;
         };
-        constructor(config: StoreConfig);
+        constructor(config: StoreConfig<T>);
         _connected(): void;
-        _disconnected(): void;
+        _disconnected(event: CloseEvent): void;
         _mutate(message: any): void;
         commit<T>(name: string, ...data: any): Promise<T | undefined>;
         subscribe(listener: (state: T) => void): () => void;
-        connection(listener: (state: boolean) => void): () => void;
     }
     export default Storex;
 }