+++ test/fixtures/stores/key_init.ex
@@ -1,11 +1,11 @@
-defmodule StorexTest.Store.Counter do
+defmodule StorexTest.Store.KeyInit do
   use Storex.Store
 
   def init(_session, _params) do
     {:ok,
      %{
        counter: 0
-     }}
+     }, "user_id"}
   end
 
   def mutation("increase", _data, _session_id, _params, state) do