+++ test/storex/browser/plug_test.exs
@@ -46,9 +46,7 @@ defmodule StorexTest.Browser.Plug do
       session
       |> visit("http://localhost:#{@port}/")
 
-    session_id = session |> text(css(".session"))
-
-    Storex.mutate(session_id, "StorexTest.Store.Counter", "increase")
+    Storex.mutate("StorexTest.Store.Counter", "increase", [])
 
     session
     |> assert_has(css(".counter-value", text: "1"))
@@ -67,5 +65,17 @@ defmodule StorexTest.Browser.Plug do
     |> fill_in(css(".input-text"), with: "John Doe")
     |> click(css(".text-send"))
     |> assert_has(css(".text-value", text: "John Doe"))
+  end
+
+  test "test join error", %{session: session} do
+    session
+    |> visit("http://localhost:#{@port}/")
+    |> assert_has(css(".error-message", text: "Unauthorized"))
+  end
+
+  def sleep(session, time \\ 10000) do
+    Process.sleep(time)
+
+    session
   end
 end