+++ test/test_helper.exs
@@ -1,2 +1,6 @@
+:ok = LocalCluster.start()
+
 Application.ensure_all_started(:wallaby)
+Application.ensure_all_started(:storex)
+
 ExUnit.start(timeout: 600_000, seed: 0)